change minion speed

+++ Assets/Scripts/Minions/MinionData.cs
@@ -5,11 +5,11 @@ namespace Minions
         public int Health;
         public int AttackDamage;
         public long AttackCdInSeconds;
-        public int Velocity;
+        public float Velocity;
         public float Range;
         public int Price;
 
-        public MinionData(int health, int attackDamage, long attackCdInSeconds, int velocity, float range, int price)
+        public MinionData(int health, int attackDamage, long attackCdInSeconds, float velocity, float range, int price)
         {
             Health = health;
             AttackDamage = attackDamage;
